[ADD]: Buttons able to clock in

+++ timetracking/src/main/java/com/nunegal/timetracking/repository/WorkingTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WorkingTypeRepository extends JpaRepository<WorkingType, Long> {
+public interface WorkingTypeRepository extends JpaRepository<WorkingType, Integer> {
 }
